Functions

MegTools.average_across_trials — Method

average_across_trials(all_epochs)

Returns the averaged signal across trials of a single condition (e.g. Sub09["1"] ).

MegTools.average_across_trials — Method

average_across_trials(all_epochs::Dict)

Returns the averaged signal across trials of a subject (Dict with multiple conditions)).

MegTools.baseline_correction — Method

baseline_correction(data; baseline_range=(-200,0), output_baseline=false)

Baseline correct data (single subject—single channel or multiple channels) based on the specified range (default is -200 ≤ t ≤ 0)

Returns baseline corrected data (single channel or multiple channels) and if output_baseline is set to true then it also returns the individual baseline

MegTools.baseline_correction — Method

baseline_correction(data::Dict; baseline_range=(-200,0))

Baseline correct data (single subject with one or more conditions) based on the input baseline values for each channel. This is usually an averaged computed baseline from many conditions. For example, by using the get_averaged_baseline function

Returns baseline corrected data (single channel or multiple channels)

MegTools.baseline_correction — Method

baseline_correction(data::Dict; baseline_range=(-200,0), output_baseline=false)

Data contains all conditions and is a Dict (subject). Baseline correct data (single channel or multiple channels) based on the specified range (default is -200 ≤ t ≤ 0)

Returns baseline corrected data (single channel or multiple channels) of all conditions and if output_baseline is set to true then it also returns the individual baselines

MegTools.collect_mean_amps — Method

collect_mean_amps(peaks::Dict, cond_trigger_vals=load_trigger_values("regsoi"))

Collects the mean amplitudes of the left and right ERF of all conditions present in the subject (Dict) input. It also converts soi triggers to values of the sois; as assigned by the load_trigger_values(experimental_paradigm) function. By default it loads the regsoi trigger values.

Returns in the following format: soi, left_mean_amps, right_mean_amps

MegTools.collect_peaks — Method

collect_peaks(peaks::Dict, cond_trigger_vals=load_trigger_values("regsoi"))

Collects the peaks (the maximum value and their latencies) of the left and right ERF of all conditions present in the subject (Dict) input. It also converts soi triggers to values of the sois; as assigned by the load_trigger_values(experimental_paradigm) function. By default it loads the regsoi trigger values.

Returns in the following format: soi, left_amps, left_lats, right_amps, right_lats

MegTools.find_mean_amplitude — Method

find_mean_amplitude(data, left_hem_channels, right_hem_channels, mean_range=(50,150))

Finds the channels containing peak values (for left and right hemisphere data sets) in (ideally averaged) data. Channels of interest are passed into lefthemchannels and righthemchannels as Symbols. The latency window for evaluating the peak values can be set with mean_range (default is set to 50 ≤ t ≤150)

Returns the left and right mean amplitudes, peak erfs and their respective channel labels in the following format: left_peak_erf, left_peak_value, left_peak_latency, right_peak_erf, right_peak_value, right_peak_latency, peak_channel_left, peak_channel_right

MegTools.find_mean_amplitude — Method

find_mean_amplitude(data::Dict, left_hem_channels, right_hem_channels, mean_range=(50,150))

Data contains all conditions and is a Dict (subject). Finds the channels containing peak values (for left and right hemisphere data sets) in (ideally averaged) data. Channels of interest are passed into lefthemchannels and righthemchannels as Symbols. The latency window for evaluating the peak values can be set with mean_range (default is set to 50 ≤ t ≤150)

Returns the left and right peak erfs and their respective channel labels as a Dict with the following entires: ["left_peak_erf"],["left_peak_value"], ["left_peak_latency"], ["right_peak_erf"], ["right_peak_value"], ["left_channel_label"], ["right_peak_latency"], ["right_channel_label"]

MegTools.find_peaks — Function

find_peaks(data, left_hem_channels, right_hem_channels, peak_range=(50,150))

Finds the channels containing peak values (for left and right hemisphere data sets) in (ideally averaged) data. Channels of interest are passed into lefthemchannels and righthemchannels as Symbols. The latency window for evaluating the peak values can be set with peak_range (default is set to 50 ≤ t ≤150)

Returns the left and right peak erfs and their respective channel labels in the following format: left_peak_erf, left_peak_value, left_peak_latency, right_peak_erf, right_peak_value, right_peak_latency, peak_channel_left, peak_channel_right

MegTools.find_peaks — Method

find_peaks(data::Dict, left_hem_channels, right_hem_channels, peak_range=(50,150))

Data contains all conditions and is a Dict (subject). Finds the channels containing peak values (for left and right hemisphere data sets) in (ideally averaged) data. Channels of interest are passed into lefthemchannels and righthemchannels as Symbols. The latency window for evaluating the peak values can be set with peak_range (default is set to 50 ≤ t ≤150)

Returns the left and right peak erfs and their respective channel labels as a Dict with the following entires: ["left_peak_erf"],["left_peak_value"], ["left_peak_latency"], ["right_peak_erf"], ["right_peak_value"], ["left_channel_label"], ["right_peak_latency"], ["right_channel_label"]

MegTools.get_averaged_baseline — Method

baseline_correction(data::Dict; baseline_range=(-200,0))

Baseline correction based on data from specified conditions.Data contains all conditions and is a Dict (subject). Baseline correct data (single channel or multiple channels) based on the specified range (default is -200 ≤ t ≤ 0)

Returns

MegTools.highlow_butterworth_filter — Method

highlow_butterworth_filter(data,sampling_rate; low_pass=30, high_pass=1, bw_n_pole=5, offset=true)

Applies a high and low-pass filter of butterworth design (n pole 5). For altering the threshold values for filters, change add keyword arguments lowpass for low pass filter cut-off (default=30) and highpass for high-pass cut-off (default=1). To change the nth order of the butterworth filter, use keyword bwnpole (default=5). Offset due to filtering at 0z, i.e the mean of the time series, can be turned off by setting offset=false which manually adds back the offset

Returns filtered data

MegTools.highlow_butterworth_filter — Method

highlow_butterworth_filter(data::Dict, sampling_rate; low_pass=30, high_pass=1, bw_n_pole=5)

Data contains all conditions and is a Dict (subject). Applies a high and low-pass filter of butterworth design (n pole 5). For altering the threshold values for filters, change add keyword arguments lowpass for low pass filter cut-off (default=30) and highpass for high-pass cut-off (default=1). To change the nth order of the butterworth filter, use keyword bwnpole (default=5).

Returns filtered data as a Dict

MegTools.load_BSepochs — Method

load_BSepochs(subject_path::String)

Reads epoch data (trials) from the outputs generated by Brainstorm and bundles them together. The general output of Brainstorm consists of .mat data files for each different epoch (trialnumber), in each different condition (datacondition_label). An example file name is data_15_trial045.mat. The output is a Dict type with labels for each condition, with data being present as KeyedArrays.

For more information on Brainstorm: Tadel F, Baillet S, Mosher JC, Pantazis D, Leahy RM (2011). Brainstorm: A User-Friendly Application for MEG/EEG Analysis Computational Intelligence and Neuroscience, vol. 2011, ID 879716. Website: https://neuroimage.usc.edu/brainstorm/

Examples

julia> load_BSepochs("reg_soi_ext1_ya10_sss")
Dict{Any,Any} with 19 entries:
  "4"  => [showing 3 of 100 slices]…
  "1"  => [showing 3 of 100 slices]…
  "12" => [showing 3 of 100 slices]…
  "2"  => [showing 3 of 100 slices]…
  ⋮    => ⋮

MegTools.load_besa_av — Method

load_besa_av(file_name)

Produces a dictionary of arrays (using AxisKeys and NamedDims), based on the input .mat file of a single average. The input .mat file is produced from the BESA®-MATLAB® interface and then saved as a separate .mat file containing the average of a single condition.

Each array has the following format: [time(ms), channels]

MegTools.load_cont_epochs — Method

load_cont_epochs(file_name)

Produces a dictionary of arrays (using AxisKeys and NamedDims), based on the different conditions classified from the input .mat file. The input .mat file is produced from the BESA®-MATLAB® interface when the "Epochs around triggers" option is selected.

Each array has the following format: [time(ms), channels, trials]

MegTools.select_channels — Method

select_channels(data; paradigm)

Select specific channels of interest to be processed for analysis. It selects channel names based on the available paradigms and returns data from channels, as well as the labels of the channels. The default paradigm is set to auditoryN1m. For custom selection use the paradigm custom_channels and input the channels as positional arguments as arrays of strings, or as a single channel string e.g.

averagefilteredauditory, auditoryleft, auditoryright = selectchannels( average, paradigm="customchannels", leftchannels=["MEG0241", "MEG0231"], rightchannels=["MEG2641",] )

Returns in the following format: selected_data, left_hem_channels, right_hem_channels

MegTools.select_channels — Method

select_channels(data::Dict; paradigm)

Data contains all conditions and is a Dict (subject). Select specific channels of interest to be processed for analysis. It selects channel names based on the available paradigms and returns data from channels, as well as the labels of the channels. The default paradigm is set to auditoryN1m.

Returns in the following format: selected_data, left_hem_channels, right_hem_channels