Functions

MegTools.average_across_trials — Method

average_across_trials(all_epochs)

Returns the averaged signal across trials of a single condition (e.g. Sub09["1"] ).

MegTools.average_across_trials — Method

average_across_trials(all_epochs::Dict)

Returns the averaged signal across trials of a subject (Dict with multiple conditions)).

MegTools.baseline_correction — Method

baseline_correction(data; baseline_range=(-200,0))

Baseline correct data (single channel or multiple channels) based on the specified range (default is -200 < t < 0)

Returns baseline corrected data (single channel or multiple channels)

MegTools.baseline_correction — Method

baseline_correction(data::Dict; baseline_range=(-200,0))

Data contains all conditions and is a Dict (subject). Baseline correct data (single channel or multiple channels) based on the specified range (default is -200 < t < 0)

Returns baseline corrected data (single channel or multiple channels) of all conditions

MegTools.collect_peaks — Function

collect_peaks(peaks::Dict, cond_trigger_vals=load_trigger_values("regsoi"))

Collects the peaks (the maximum value) of the left and right ERF of all conditions present in the subject (Dict) input. It also converts soi triggers to values of the sois; as assigned by the load_trigger_values(experimental_paradigm) function. By default it loads the regsoi trigger values.

Returns in the following format: soi, left_amps, right_amps

MegTools.find_peaks — Function

find_peaks(data, left_hem_channels, right_hem_channels, peak_range=(50,150))

Finds the channels containing peak values (for left and right hemisphere data sets) in (ideally averaged) data. Channels of interest are passed into lefthemchannels and righthemchannels as Symbols. The latency window for evaluating the peak values can be set with peak_range (default is set to 50 < t <150)

Returns the left and right peak erfs and their respective channel labels in the following format: left_peak_erf right_peak_erf, peak_channel_left, peak_channel_right

MegTools.find_peaks — Function

find_peaks(data::Dict, left_hem_channels, right_hem_channels, peak_range=(50,150))

Data contains all conditions and is a Dict (subject). Finds the channels containing peak values (for left and right hemisphere data sets) in (ideally averaged) data. Channels of interest are passed into lefthemchannels and righthemchannels as Symbols. The latency window for evaluating the peak values can be set with peak_range (default is set to 50 < t <150)

Returns the left and right peak erfs and their respective channel labels as a Dict with the following entires: ["left_peak_erf"],["left_peak_value"], ["right_peak_erf"], ["right_peak_value"], ["left_channel_label"], ["right_channel_label"]

MegTools.highlow_butterworth_filter — Method

highlow_butterworth_filter(data,sampling_rate; low_pass=30, high_pass=1, bw_n_pole=5)

Applies a high and low-pass filter of butterworth design (n pole 5). For altering the threshold values for filters, change add keyword arguments lowpass for low pass filter cut-off (default=30) and highpass for high-pass cut-off (default=1). To change the nth order of the butterworth filter, use keyword bwnpole (default=5).

Returns filtered data

MegTools.highlow_butterworth_filter — Method

highlow_butterworth_filter(data::Dict, sampling_rate; low_pass=30, high_pass=1, bw_n_pole=5)

Data contains all conditions and is a Dict (subject). Applies a high and low-pass filter of butterworth design (n pole 5). For altering the threshold values for filters, change add keyword arguments lowpass for low pass filter cut-off (default=30) and highpass for high-pass cut-off (default=1). To change the nth order of the butterworth filter, use keyword bwnpole (default=5).

Returns filtered data as a Dict

MegTools.load_cont_epochs — Method

load_cont_epochs(file_name)

Produces a dictionary of arrays (using AxisRange and NamedDims), based on the different conditions classified from the input .mat file. The input .mat file is produced from the BESA®-MATLAB® interface when the "Epochs around triggers" option is selected.

Each array has the following format: [time(ms), channels, trials]

MegTools.select_channels — Method

select_channels(data; paradigm)

Select specific channels of interest to be processed for analysis. It selects channel names based on the available paradigms and returns data from channels, as well as the labels of the channels. The default paradigm is set to auditoryN1m.

Returns in the following format: selected_data, left_hem_channels, right_hem_channels

MegTools.select_channels — Method

select_channels(data::Dict; paradigm)

Data contains all conditions and is a Dict (subject). Select specific channels of interest to be processed for analysis. It selects channel names based on the available paradigms and returns data from channels, as well as the labels of the channels. The default paradigm is set to auditoryN1m.

Returns in the following format: selected_data, left_hem_channels, right_hem_channels